pyproject.toml
[tool.poetry]
name = "am-social-media-analytics"
version = "0.1.0"
description = "Application of Social Media Analytics to Identify Business Models and Use Cases"
authors = ["Abubakar Shola Suleiman <abubakar.suleiman@tuhh.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
jupyter = "^1.0.0"
pandas = "^2.2.1"
matplotlib = "^3.8.4"
selenium = "^4.19.0"
requests = "^2.31.0"
webdriver-manager = "^4.0.1"
beautifulsoup4 = "^4.12.3"
linkedin-scraper = "^2.11.2"
python-dotenv = "^1.0.1"
seaborn = "^0.13.2"
nltk = "^3.8.1"
prefect = {extras = ["email"], version = "^3.1.10"}
pyarrow = "^18.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
